rework the project and add add rentalplan entitie

+++ src/FleetManager.Domain/Repositories/ToRental/IRentalUpdateOnlyRepository.cs
@@ -1,10 +1,9 @@
-using System;
 using FleetManager.Domain.Entities;
 
 namespace FleetManager.Domain.Repositories.ToRental;
 
 public interface IRentalUpdateOnlyRepository
 {
-    Task<Rental?> GetById(long id);
+    Task<Rental?> GetById(int id);
     void Update(Rental rental);
 }
